Rename root project plot -> output
