+++ output/config/gnuplot.yaml
