+++ output/config/plots.yaml
