+++ output/launch/generate_plots.launch
