+++ output/src/aggregate.cpp
@@ -1,10 +1,10 @@
 #include <cmath>
 
-#include <moveit_benchmark_suite/plot/aggregate.h>
+#include <moveit_benchmark_suite/output/aggregate.h>
 #include <moveit_benchmark_suite/io.h>
 
 using namespace moveit_benchmark_suite;
-using namespace moveit_benchmark_suite::plot;
+using namespace moveit_benchmark_suite::output;
 
 //
 // statistic
