+++ output/src/generate_plots.cpp
@@ -1,11 +1,11 @@
 
 #include <ros/ros.h>
-#include <moveit_benchmark_suite/plot/gnuplot.h>
-#include <moveit_benchmark_suite/plot/htmlplot.h>
+#include <moveit_benchmark_suite/output/gnuplot.h>
+#include <moveit_benchmark_suite/output/htmlplot.h>
 #include <moveit_benchmark_suite/benchmark.h>
 
 using namespace moveit_benchmark_suite;
-using namespace moveit_benchmark_suite::plot;
+using namespace moveit_benchmark_suite::output;
 
 constexpr char INPUT_PARAMETER[] = "input_files";
