+++ output/src/plot_dataset.cpp
@@ -1,10 +1,10 @@
 
 #include <ros/ros.h>
 #include <moveit_benchmark_suite/profiler.h>
-#include <moveit_benchmark_suite/plot/gnuplot.h>
+#include <moveit_benchmark_suite/output/gnuplot.h>
 
 using namespace moveit_benchmark_suite;
-using namespace moveit_benchmark_suite::plot;
+using namespace moveit_benchmark_suite::output;
 
 constexpr char INPUT_PARAMETER[] = "input_files";
